fix(attention): address review findings on mxfp8_attention_sm120
- binding: check cudaFuncSetAttribute/launch errors; reject cross-device tensors
- python: place LPT work lists on the plan's device (multi-GPU fix); drop unused unpack
- tile_scheduler: fix 10-of-11-member no-work WorkTileInfo initializers (latent
  counter/ptr shift); SM120-specific include guard; move into mxfp8_attention_sm120
  namespace (was top-level flashinfer, colliding with the Hopper copies)
- remove dead named_barrier.cuh (unused Hopper leftover, duplicate include guard)
- kernel: null-guard out_Mnb independently of out_Ppre; brace else+unroll sites;
  drop unused host includes
- trace: declare out=bfloat16/lse=float32 in templates, regenerate fi_trace_out
- tests: add zero-length (qo_len==0) request case

8/8 attention tests pass on RTX 5060 Ti; ruff + clang-format clean

+++ flashinfer/trace/templates/mxfp8_attention_sm120.py
@@ -43,8 +43,8 @@ mxfp8_attention_sm120_fwd_trace = TraceTemplate(
         "causal": Scalar("bool"),
     },
     outputs={
-        "out": Tensor(["total_q", "num_qo_heads", "head_dim"]),
-        "lse": Tensor(["total_q", "num_qo_heads"]),
+        "out": Tensor(["total_q", "num_qo_heads", "head_dim"], dtype="bfloat16"),
+        "lse": Tensor(["total_q", "num_qo_heads"], dtype="float32"),
     },
     constraints=[
         "num_qo_heads % num_kv_heads == 0",
@@ -79,8 +79,8 @@ mxfp8_attention_sm120_run_trace = TraceTemplate(
         "v_scale": Scalar("float"),
     },
     outputs={
-        "out": Tensor(["total_q", "num_qo_heads", "head_dim"]),
-        "lse": Tensor(["total_q", "num_qo_heads"]),
+        "out": Tensor(["total_q", "num_qo_heads", "head_dim"], dtype="bfloat16"),
+        "lse": Tensor(["total_q", "num_qo_heads"], dtype="float32"),
     },
     constraints=[
         "num_qo_heads % num_kv_heads == 0",
